requirements.txt
GitPython==0.3.6
requests==2.5.1
